refactor: Move IMessage and dependant types to @chainlit/components
