+++ frontend/src/components/organisms/chat/message/message.tsx
@@ -4,13 +4,12 @@ import { useRecoilValue } from 'recoil';
 
 import { Box, Stack } from '@mui/material';
 
-import { IMessageElement } from '@chainlit/components';
+import { IMessageElement, INestedMessage } from '@chainlit/components';
 
 import { highlightMessage } from 'state/chat';
 import { settingsState } from 'state/settings';
 
 import { IAction } from 'types/action';
-import { INestedMessage } from 'types/chat';
 
 import AskUploadButton from './AskUploadButton';
 import Author, { authorBoxWidth } from './author';
