+++ frontend/src/components/socket.tsx
@@ -9,7 +9,7 @@ import {
 } from 'recoil';
 import io from 'socket.io-client';
 
-import { IElement, TFormInput } from '@chainlit/components';
+import { IElement, IMessage, TFormInput } from '@chainlit/components';
 
 import { useAuth } from 'hooks/auth';
 
@@ -29,7 +29,7 @@ import { projectSettingsState } from 'state/project';
 import { sessionIdState, userEnvState } from 'state/user';
 
 import { IAction } from 'types/action';
-import { IMessage, IMessageUpdate, IToken } from 'types/chat';
+import { IMessageUpdate, IToken } from 'types/chat';
 
 const compareMessageIds = (a: IMessage, b: IMessage) => {
   if (a.id && b.id) return a.id === b.id;
