+++ frontend/src/types/chat.ts
@@ -1,31 +1,9 @@
 import { Socket } from 'socket.io-client';
 
+import { IMessage } from '@chainlit/components';
+
 import { IMessageElement } from './element';
 import { IMember } from './user';
-
-interface IBaseTemplate {
-  template?: string;
-  formatted?: string;
-  template_format: string;
-}
-
-export type PromptMessageRole = 'system' | 'assistant' | 'user' | 'function';
-
-export interface IPromptMessage extends IBaseTemplate {
-  role: PromptMessageRole;
-  name?: string;
-}
-
-export type ILLMSettings = Record<string, string | string[] | number | boolean>;
-
-export interface IPrompt extends IBaseTemplate {
-  provider: string;
-  id?: string;
-  inputs?: Record<string, string>;
-  completion?: string;
-  settings?: ILLMSettings;
-  messages?: IPromptMessage[];
-}
 
 export interface IChat {
   id: number;
@@ -33,22 +11,6 @@ export interface IChat {
   author?: IMember;
   messages: IMessage[];
   elements: IMessageElement[];
-}
-
-export interface IMessage {
-  id: string;
-  author: string;
-  authorIsUser?: boolean;
-  waitForAnswer?: boolean;
-  content?: string;
-  createdAt: number | string;
-  humanFeedback?: number;
-  language?: string;
-  indent?: number;
-  parentId?: string;
-  isError?: boolean;
-  prompt?: IPrompt;
-  streaming?: boolean;
 }
 
 export interface IMessageUpdate extends IMessage {
@@ -59,10 +21,6 @@ export interface IToken {
   id: number | string;
   token: string;
   isSequence: boolean;
-}
-
-export interface INestedMessage extends IMessage {
-  subMessages?: IMessage[];
 }
 
 export interface FileSpec {
