+++ libs/components/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './element';
 export * from './Input';
+export * from './messages';
 export * from './NotificationCount';
